+++ app/builders/dc20_character/classes/bard_builder.rb
@@ -8,7 +8,6 @@ module Dc20Character
       def call(result:)
         result[:combat_expertise] = COMBAT_EXPERTISE
         result[:health] = { current: 7, temp: 0 }
-        result[:path] = ['spellcaster']
 
         result
       end