+++ app/builders/dc20_character/classes/wizard_builder.rb
@@ -8,8 +8,6 @@ module Dc20Character
       def call(result:)
         result[:combat_expertise] = COMBAT_EXPERTISE
         result[:health] = { current: 8, temp: 0 }
-        result[:spell_list] = ['arcane']
-        result[:path] = ['spellcaster']
 
         result
       end