+++ app/decorators_v2/dc20/classes/druid_decorator.rb
@@ -3,14 +3,18 @@
 module Dc20
   module Classes
     class DruidDecorator < ApplicationDecoratorV2
-      def call(result:)
+      def call(result:) # rubocop: disable Metrics/AbcSize
         @result = result
         @result['max_stamina_points'] = paths['martial']
         @result['max_mana_points'] = mana_points_by_level
         @result['maneuver_points'] = paths['martial']
         @result['max_health'] = 6 + level + modified_abilities['mig']
         @result['spells'] = spells_by_level
-        @result['spell_lists_amount'] = 1
+        @result['spell_list'] = ['primal']
+
+        @result['spell_class'] = 'druid'
+        @result['spell_filter'] = { 'source' => 'primal' }
+
         @result
       end